+++ module/settings/settings-helpers.mjs
@@ -1,6 +1,8 @@
 import { HEROSYS } from "../herosystem6e.mjs";
 import { CreateHeroCompendiums } from "../heroCompendiums.mjs";
 
+export let overrideCanAct = false;
+
 export default class SettingsHelpers {
     // Initialize System Settings after the Init Hook
     static initLevelSettings() {
@@ -282,7 +284,37 @@ export default class SettingsHelpers {
             default: "1.0.0",
             requiresReload: true,
         });
+
+        // Keybinding for Override
+        game.keybindings.register(module, "OverrideCanAct", {
+            name: game.i18n.localize("keybindings.OverrideCanAct.Name"),
+            hint: game.i18n.localize("keybindings.OverrideCanAct.Hint"),
+            onDown: handleOverrideCanAct,
+            onUp: handleOverrideCanAct,
+            editable: [
+                {
+                    key: "ControlLeft",
+                },
+            ],
+            precedence: -1,
+        });
     }
+}
+
+function handleOverrideCanAct(event) {
+    // When a dialog box (like spend STUN for END) shows it call the event.up for some unknown reason.
+    // These extra/unwanted events are not trusted.
+    if (!event.event.isTrusted) {
+        //console.log("untrusted", event);
+        return;
+    }
+
+    if (event.up) {
+        overrideCanAct = false;
+    } else {
+        overrideCanAct = event.key;
+    }
+    //console.log(overrideCanAct, event);
 }
 
 class AutomationMenu extends FormApplication {